Crowd-sourced projects (#70)
* feat: New crowd-sourced field for projects

First in a series of commits to introduce crowd-sourced support.

This first commit only introduces a new DB field in the projects table
that will store a boolean indicating if the project is crowd-sourced.

There is a lot of code churn here, but this is all in src/tests as there
are a lot of calls to createProject which has had a change to it's
parameters.

I've also addressed a security hole I found in working on this:
User X could submit API calls making changes to User Y's projects if
User X and User Y were both in the same class.

Not a drastic security hole, and I doubt it was ever exploited. But it
was worth addressing. I've added that in to verifyProjectAccess as that
should pick it up for all relevant APIs in one.

Contributes to #23

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>

* fix: Address minor security hole

GET requests to REST APIs would allow a student to fetch info about
another student's project.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>

* chore: Additional info in Visual Recognition Slack notifications

Unrelated to this branch, but I needed some additional debug while I
test the crowd-sourcing work.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>

* feat: Support for fetching projects including crowdsourced

The GET for /projects will now return crowdsourced projects in the list.

Contributes to #23

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>

* feat: Handle permission for crowd-sourced projects on GET /projects/projid

Contributes to #23

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>

* feat: Auth updates for all REST APIs

This commit introduces two levels of auth checking.

1) Project access - For project owners, and students accessing
crowd-sourced projects in their class
2) Project owner - For project owners only

Project access is mostly read-only, except for adding and deleting
training items.

Contributes to #23

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>

* feat: UI support for crowd-sourced projects

This modifies most pages in the UI to remove controls from students
accessing crowd-sourced projects.

I needed to change the URL structure to provide user-ids rather than
the previous approach of taking it from the logged in user profile.

Closes #23

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>