doc: Improvements to the Help page
I improved the UI of the Help page a little.

I grouped the FAQs by theme to make them a little easier to find.

I added some additional FAQs to describe the information that is stored
about users, and to explain that class accounts can't be deleted through
the UI at the moment.

Contributes to: #98

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>