fix: Handle long classifier IDs
Visual Recognition generates a classifier ID by appending a timestamp to
the provided project name.

Project names can be 36 characters long, which means classifier IDs can
potentially be 46 characters long.

They are stored in a DB column that was 36 characters long. This led to
an error in production this afternoon:

Unexpected failure to train image classifier for project :
 Data too long for column 'classifierid' at row 1

This commit modifies the columns where classifier IDs are stored to
allow these longer IDs to be stored.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>