feat: Prevent Scratch extensions from polling status when user inactive
It is common for Scratch instances to be left running for hours at a
time, calling the /status API every 30 seconds.

This commit monitors whether the user is still active (as determined by
any mouse or keyboard activity). After 15 minutes with no mouse and no
keyboard activity, the status polling will stop.

Closes: #103

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>