feat: Handle unmanaged Bluemix classifiers
This commit introduces two new features:

1) Automatically deleting unknown classifiers in managed classes
For tenants where the teacher does not have the Bluemix API keys, there
should never be a classifier that is not in the DB.
For these tenants, the classifiers are automatically deleted when found
by the scheduled job.

2) Notifying teachers about unknown classifiers in unmanaged classes
For tenants where the teacher provided their own Bluemix API keys, there
is a chance that they created classifiers outside of MLforKids.
For these tenants, there is a new API to check for this, and to delete
them if wanted.
There is also a new section of the Teachers Admin page to display the
unknown classifier details.

Closes: #110

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>