feat: Cache Watson API calls from Scratch Keys
Classifications returned to Scratch now include the timestamp that the
classifier used to generate the classification was trained.

Calls made to the /classify endpoint using Scratch Keys now support the
If-Modified-Since header. When Scratch provides this header, it is
compared with the time that a classifier was last trained. If the
timestamps match, we can safely bypass the need to submit the API call
to Watson, and return a Not-Modified response instead.

This also introduces caching to Scratch for images for the first time.
I'm using md5 to create a key for the cache.

Contributes to: #119

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>