feat: Support for load balancing
The app will now be run in multiple regions, with a load balancer
in front. There is a risk that two instances started at the same
time could clash if running the cron jobs simultaneously, so I'm
introducing a new environment variable (set in manifest files) to
identify the primary region.

Contributes to: #117

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>