fix: Fix stalled training UI
I'm not able to recreate this reported problem, but examining the code
shows some mixing of timer approaches which might be a possible cause,
so this commit fixes that.

I've also added some notes to the Help page to encourage people to
refresh the page if they see no updates for ages.

Closes: #118
Closes: #112

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>