fix: Workaround for deleted VR classifiers reappearing
This has been reported by many users on many occassions, but I've only
just been able to recreate it and work out the cause.

When the service gets busy, this scenario starts becoming common:

1) user starts training classifier (attempt-1)
2) user cancels the training by deleting the attempt-1 classifier
3) cancelled/deleted classifier disappears from their list of classifiers
4) this lets them start training a new classifier (attempt-2)
5) a few minutes later, the attempt-1 classifier has finished training and
   appears in their list of classifiers as ready

The user now has two classifiers: attempt-1 and attempt-2. Which is more than
a free tier API key is allowed. The first attempt-1 classifier is divorced /
orphaned from Machine Learning for Kids. And even if they delete the second
attempt-2 classifier, they still are at their limit so they can't train any
new classifiers.

This commit adds a horrid kludgey workaround for this frustrating
behaviour.

When we cancel the training of a classifier, we'll wait fifteen minutes
(assuming that most classifiers will complete within this time). And then we'll
delete it again, so if it's reappeared, we'll be able to clean up before
the user notices a problem.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>