feat: Support for new IAM-auth mechanism in Watson Assistant (#124)
Watson Assistant uses IAM-based auth (using short-lived bearer tokens in
request headers) in US-East and Sydney. And basic-auth with a username
and password in all the other regions.

This commit adds the back-end support for managing the new type of
API-key, and using it in the HTTP requests to the Assistant API.

I've still not made the UI changes, so noone will be able to store the
new-style credentials yet. But if they could, they'd work!

Contributes to: #114

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>