chore: Add Travis for verifying code compiles and lints cleanly
As part of the broader move to Grunt, this introduces a new Gruntfile.
To start with, this is being used for compile/lint only, but in future
this will eventually replace the current gulp tasks.

As much logic is kept out of Travis yml as possible, with Travis just
being used to kick off the task runner - so that we can ensure that
local compile/lint is done in exactly the same way as Travis
compile/lint.

Closes: #6

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>