chore: Make environment variable names into constants
There's a very small benefit from doing this, that the TypeScript compiler
can catch if you put a typo in an environment variable name. Only slightly
not pointless then.

I've added all the required variables to the startup check, so the app will
refuse to start if it's missing any required environment.

Contributes to: #4

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>