chore: Shutdown on SIGTERM
This should slightly speed up some Cloud Foundry admin actions, such
as stops and restarts, as we won't need to wait for CF to aggressively
kill app instances. By respecting and handling the SIGTERM signal
that CF sends (and waits 10 seconds for) on a stop, this will save a
little time.

Closes: #33

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>