chore: Shutdown on SIGTERM (#35)
* chore: Make environment variable names into constants
There's a very small benefit from doing this, that the TypeScript compiler
can catch if you put a typo in an environment variable name. Only slightly
not pointless then.
I've added all the required variables to the startup check, so the app will
refuse to start if it's missing any required environment.

Contributes to: #4

* chore: Shutdown on SIGTERM
This should slightly speed up some Cloud Foundry admin actions, such
as stops and restarts, as we won't need to wait for CF to aggressively
kill app instances. By respecting and handling the SIGTERM signal
that CF sends (and waits 10 seconds for) on a stop, this will save a
little time.

Closes: #33

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>