fix: Avoid false-positives about Visual Recognition credentials
Visual Recognition occassionally throws an HTTP-500 with a "please try
again" error. An email was sent to a teacher today saying that their
credentials don't work, because of this. Their credentials are probably
fine. This commit adds a small check for this error type, to prevent
sending future emails like this.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>