feat: Improve performance with gzip for static resources
As CloudFlare is caching all the static resources now, I don't need to
be too nervous about the CPU cost of doing gzip compression on the fly
in a production site. In practice, it will only happen for the first few
requests, and then CloudFlare will just handle hosting the
already-gzipped resources.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>