fix: Limit the amount of images downloaded
Previously, training an image classifier could mean downloading up to
200 images of unlimited size.

With this commit, training an image classifier will mean downloading at
most 100 images, with each image no bigger than 5mb each.

This still would be enough to crash the deployed app (as we're not
running with 500mb disk space) so this needs to be followed with other
steps. But it is a start.

Contributes to: #41

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>