fix: Reduce size of images used for training
Assume a student has 100 images in their project, and each one is about
5mb (because they're dragging in the biggest high-res images from an
images web search).

Downloading them, as part of preparing a training zip for Visual
Recognition, will need 500mb of disk space. The fact that we'll put them
in a zip doesn't really help, as they won't get much smaller in a zip -
being already compressed. In fact, in putting it in a zip, we need more
disk space - as temporarily we'll have two copies.

Now assume several students are doing this at once.

This isn't sustainable.

This commit moves this in a more sensible direction. When the images are
downloaded, they're not downloaded directly to disk. Rather they are
piped to a resize filter that resizes the images aggressively to 800
pixels across. And then that is put in the zip.

I'm reasonably confident that this won't noticeably impact the machine
learning models that are trained, and I think it'll make it possible to
run training for many more concurrent students.

There is a risk that I'm going to significantly increase memory usage,
to perform the resizing. That's something I'll need to keep an eye on.
But a little testing so far looks like it should be okay.

Contributes to: #41

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>