chore: Update new-account email template
The old email template included a date-range for when the new account
was needed, because I needed to avoid hitting the limit on Auth0's free
tier - it meant I could identify classes that were safe to delete.

Now that I'm on the Auth0 OSS program, and the limit is much more
generous, I don't need to delete inactive accounts.

This commit removes the line in the email template, and the notes in the
web page.

Contributes to: #28

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>