fix: Protect against project names that cannot be stored
Project names are stored using MySQL's 'utf8' data type, which isn't
true UTF-8, but a smaller subset that doesn't include longer (in bytes)
characters.

This meant longer UTF-8 characters like emoji cannot be stored.

The two main approaches to fixing this are either to switch to a
different MySQL type (e.g. utf8mb4) or to use input validation to
restrict the new-project requests to a tolerable subset.

I've gone with the latter. It's enough of a character range to include
accented characters, so I don't think it will cause any multilingual
problems. And I don't mind not allowing projects to be named with
emojis.

Closes: #58

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>