fix: Quiz UI sometimes not reflecting choices
There seems to be a random timing bug for some browsers with the
quiz. The user clicks on the checkbox. Their choice is reflected
by a change to ng-model. And a function invoked by ng-click
checks to see if their answer is right.

The problem is that for some users, it looks like ng-click is
invoking the function before ng-model has updated the model.
So the function thinks there is no selected answer, and declares
it wrong.

A 0-ms timeout seems to be enough to force the ng-model update
to complete before checking the answer.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>