feat: Scratch 3 support
Initial implementation of a Scratch 3 extension. Only the text extension
is implemented so far, as it's the simplest of the three.

It's functional enough to be used, so it's shown in the UI - although
clearly marked as an early alpha.

Almost all of the functionality from the Scratch 2 version (e.g. caching,
throttling, support for getting labels and confidence scores, ability to
add training data, etc.) is present.

The only thing missing is the status indicator, which doesn't seem to be
included in the Scratch 3 UI. It means there's no clear way to show
whether the machine learning model is ready / functional.

Contributes to: #21

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>