feat: Scratch block improvements
Modified the label field in the add-training-data block in the
Scratch extension so that it is a list of the labels, rather
than a free text field.

The benefit of this is that it's more obvious what should go
into that field, and also acts as validation - as it means
you can only choose a valid label.

The downside is that it means you can't use list or reporter
blocks in there, which would break some of the project worksheets
which depend on that.

As a result, I've left the existing add-training-data block as
well as the new list-based one.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>