feat: Display confirmation when verifying email addresses
When users click the Verify Email Address link they get in an email, it
will go first to auth0.com, and then redirect to
machinelearningforkids.co.uk with a query parameter confirming the
outcome.

This commit looks for that query parameter and displays a dialog if
found to let the user know that it worked.

It uses window.location to send the user to the home page while clearing
the query parameter to prevent it being displayed again.

Closes: #9

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>