chore: Fix TypeScript compile errors
I'm not sure what's changed, but TypeScript is refusing to compile the
tests, reporting about 70 errors.

It's a blocker for some other work, so I just want to get the tests
compiling for now. But this needs re-doing properly - I've had to really
hack the tests to get tsc to accept them.

(To be fair, some of the errors that TypeScript reported are legitimate,
so it's not all bad for me to revisit this)

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>