feat: Keep users logged in while active
Access tokens fetched from Auth0 have an expiry time of 2 hours by
default. This is almost always good enough, as schools rarely run coding
lessons that are longer than 2 hours. However for some events where
computers are left logged in, this has caused problems as users need to
log in again while working.

This commit uses setTimeout to schedule the refreshing of the access
tokens before they expire. I've decided not to use refresh tokens, as
that would be difficult to keep secure in the browser.

Closes: #125

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>