chore: Track user-agents that are using Scratch keys
I've noticed an increase in errors from invalid requests sent to
scratchkey API endpoints. I'm hoping it's just bugs in student's Python
projects, and not a bug in the Scratch extensions. But to make sure,
I'm adding custom user agents to the Scratch extensions, so I can track
where the errors are coming from.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>