feat: Explain image ML model failures and offer fixes
Server logs this week show a large number of machine learning model
failures for image projects, as a result of failures to download
training images when preparing the zip files for submitting to the
Watson API.

This was shown in the UI as a generic "Training failed" error, with no
further feedback or info to students to help them resolve the problem.

This commit introduces a new dialog that will be displayed in such cases
- identifying the image that caused the problem, and asking the student
if they would like the image to be deleted. In most cases, after doing
this, the student will be able to try training the ML model again, and
this time be successful.

Closes: #138

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>