fix: Protect against duplicate field names in number projects
Projects could be created with multiple identically-named fields. This
breaks the training UI (as it only sends one value for each field name
and so the training data was missing values for some fields).

This commit protects against that in a couple of places - adding a
check that the UI has sent at least one value for each field, and
prevents this sort of project being created again in future.

Closes: #144

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>