fix: Minor fixes for errors seen in prod
A teacher with no API keys in their class account will generate
a server-error any time they visit the 'Supervision' page (which
generates an API call for unmanaged classifiers).

This will have been silently swallowed by the UI, but it does
result in an HTTP-500 response which I'm trying to squash so it's
good to tidy this up so it properly returns an empty list.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>