chore: Track session user usage
If someone tries to use a session-user id after it has been cleaned
up by the cron job, this is reported as an error (because the SQL
statement used to retrieve the token doesn't find anything).

I *think* these aren't actually errors, and are just the result of
someone opening the browser many days/hours after using Try It Now,
but this commit adds some extra debug/logging so I can make sure.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>