feat: Explain the recognise image blocks
Error logs suggest that students not following a pre-written
worksheet rarely understand how to use the 'recognise image'
blocks. They either leave the word 'image' in there, or put
costume names or numbers in. But they rarely realise that they
need to put image data in there.

This commit will display a pop-up with some additional help
text and graphics after someone has got it wrong a couple of
times.

Pop-up blockers are very common nowadays, so I've gone the
long way around on this to tie the pop-up to a user action.
This means I've got an annoying bit of indirection in the way
as I display a message saying I want to show them a pop-up
and force them to click to display the pop-up.

This is particularly long-winded in Scratch 3, as extensions
are run in Web Workers, and as such don't have access to
window/document/etc. so an extension can't display a pop-up.
So I've had to pass a message back to Scratch instead.

(I did spend many many hours trying to figure out how to
use the Tips Library for this, and display the Tips Library
window instead, but I couldn't make it work. I might come
back to that some day)

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>