fix: POST support for adding training images
Scratch was using query parameters in GET requests for uploading images,
because that was easiest to get working with ScratchX. Scratch 3 makes
POST requests easy, so that is no longer a restriction. And CloudFlare
has been rejecting some requests with long query parameters recently, so
switching this to a POST seems like a sensible precaution.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>