feat: Use thumbnails instead of original image from Wikimedia
Wikimedia is a common source of training images - with thousands of
training images in the current training data table. It looks like
the most common choice by students is the fullsize/original image,
often several thousand pixels in width. This puts an overhead on
ML for Kids needing to download the several-MB image file and
resize it in memory.

This commit treats wikimedia as a special case, and downloads the
thumbnail version of a wikimedia photo instead of the original.

This should significantly reduce the overhead for preparing the
training zip for wikimedia projects.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>