fix: Handle unexpected error payloads from Visual Recognition
I noticed some weird errors in production this afternoon. It showed up
as MySQL errors ("Bind parameters must not contain undefined. To pass
SQL NULL specify JS null") because I was trying to store a classifier in
the DB with a null ID and timestamp.

Looking at the trace, it's because of failed requests to train a visual
recognition classifier that returned an HTTP-200 response with this
payload:

{"error":{"code":200,"error_id":"input_error","description":"undefined error"}}

This commit is a quick-and-dirty hack to look out for this and handle it
as a training error (rather than an unexpected MySQL error).

In the meantime, I'll try and chase Visual Recognition to see why this
is happening.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>