docs: Update projects to Scratch 3
I've been refreshing the project worksheets - mostly to update them
to use Scratch 3 instead of the old ScratchX site, but also to
refine some instructions and replace screenshots that were taken
before the site got a logo and minor UI refresh.

Most of this is happening in IBM/taxinomitis-docs but this commit
is a small part of that work.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>