fix: Explain auth failures to fetch training images
A common source of errors when students drag training images into their
training buckets is that they drag an image from a site that they have
access to, but that requires authentication. In the absence of the
credentials, MLforKids gets an auth error when it tries to download the
image. This has confused students, so this commit updates the error
message to explain what is happening.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>