feat: DB and REST API support for managing site alerts
This commit adds support for storing messages to be displayed to
users. The intention is that this will be used to notify users
of outages or incidents that are causing errors.

I've left an "info" severity in, as this would let it be used
to display messages that aren't errors, but I've got no current
plans for using that.

Contributes to: #190

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>