feat: Display site alerts in a new site banner
This commit adds a new alert bar at the top of the web UI.

The site alerts will be refreshed when users log in or off, and then
periodically (currently every 5 minutes) after that.

The commit also adds a very simple UI for creating new alerts, but
this is pretty rough and could do with cleaning up.

Contributes to: #190

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>