docs: Restructure Headlines projects
There was an easy version and an advanced version of the Headlines project.
That was hard to include in the Worksheets list in the UI, as it could only
live in one place. This commit separates those versions out into two separate
projects so they can both be represented in the Worksheets list.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>