fix: Reduce request payload size for describe-model OpenWhisk functions
OpenWhisk is choking on requests over 128k. The noughts and crosses project
with 1000 training examples is coming out to about 350kb for me, so it
won't work.

The labels and field names are repeated a lot in the training data, so it
is easy to compress it. This commit has a quick and dirty compression by
replacing repeated strings with an int index into a key of strings, which
brought the 350kb noughts and crosses project training down to about 80kb.

I think I'll need something better in time, but this should adress most
use cases. I've got "beta" labels everywhere to cover the exceptions.

Contributes to: #220

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>