feat: Decision tree visualisations (#221)
A new feature for explaining how the machine learning models work for numbers
projects. It generates an interactive visualisation of a decision tree, that
highlights the nodes and edges used for predictions of any given test data.

It's using graphviz to generate an SVG from Python, hosted as an OpenWhisk
action for scalability. And then there is some browser-side JavaScript to
add/remove classes to edges and nodes in the SVG to make it interactive.

The visualisation gets pretty big, so the UI also has some basic pan and
zoom functionality.

Closes: #220 

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>