fix: Explicitly specify MySQL charset
Two developers reported problems setting up a development
environment, with the symptom being a failing test (with
MySQL not rejecting some special characters that the test
expected it to).

I'm unable to reproduce this problem, even running the same
MySQL version as them, so my best guess is that they have
a different charset/collation to me (as I've provided no
advice to create the database with anything other than
"CREATE DATABASE" they presumably inherited system defaults).

This commit explicitly sets the CHARSET and COLLATION to
what the production site DB has as it's default.

I'm hoping this will resolve the problem for them.

Contributes to: #225

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>