fix: Create very large training zips locally
The most common reason for failing to create training zip files in OpenWhisk
is that the serverless function CreateZip creates a zip file that is larger
than the 5mb response payload limit that IBM Cloud Functions allows.

This needs some significant redesign/restructuring to handle it, but for now,
we'll just workaround the problem by handling these (fortunately rare) edge
cases by building the zip outside of OpenWhisk.

This is only done for unexpected failures - if the OpenWhisk function returns
an error payload, we just use the error description in it.

Contributes to: #263

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>