feat: Remove common characters rejected by Watson Assistant
Errors today show classify requests being rejected
by Watson Assistant because there were newlines in
the input text.

It looks like these came from the Twitter extension.
It's hard for these errors to be surfaced in the
Scratch UI so it likely wouldn't have been clear to
the students why these requests failed.

It's simplest for me to clean this up for them in the
API.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>