fix: Validate credentials when getting HTTP413 from Vis Rec
When training a new ML model, Visual Recognition will sometimes return an
HTTP 413 (Request Entity Too Large) response in the event of an auth
problem - presumably because only authorised users can post large payloads.

This commit is a small change to the error handling for training new models
so that if we get an HTTP 413, before returning that, we check the
credentials (by using them to fetch a list of classifiers) to see if that
might be the root cause of the error.

Closes: #279

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>