feat: Check Watson credentials when new projects are created
When a student creates a new project, the UI will now make
a check that there are available working API keys for that
type of project.

This is only needed for text and images projects.

A warning will be displayed if there are no API keys for that
project type, or if there are API keys but they are _all_
rejected by IBM Cloud.

The warning will be displayed to students on the projects list
with advice to ask their teacher to check the API keys.

This won't be done for managed classes, as the teacher won't
be able to do anything about it.

The checks are cached fairly aggressively both in the browser
and in a server-side LRU cache, to avoid unnecessary spamming
to Watson APIs and the DB when a large class all create a new
project at approximately the same time.

Closes: #282

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>