feat: Add pre-trained models section
New section of the site, dedicated to hosting a curated set of
pre-trained ML models.

The focus is going to be in providing models that are:
- not feasible for children to train themselves, because of the
  amount of training data they'd need to be able to do it
- fun or interesting uses of ML

I've only got a couple of ML models to start with, but the plan
is to add additional ones soon.

I will also start adding some worksheets that demonstrate how
to use the new blocks.

The new section isn't linked into the rest of the site. This is
mostly because I haven't worked out how/where to do that yet,
but it will also let me soft-launch the feature by sharing the
URL with some potential testers.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>