feat: Support for teachers reviewing training data
This commit is an update to the Teacher -> Supervision page, so
that the list of student projects is now a list of links to the
training page for each project. The supervision training page is
a read-only view, to allow teachers to review the training data
not contribute to it.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>