feat: Prioritize reusing managed pool credentials after deletion
This doesn't apply to Visual Recognition which has monthly usage
limits (so in the event of a failure, it is best not to use the
credentials again until the next month).

For Assistant, to minimize the number of credentials that are used
this commit adjusts the lastfail timestamp after models are deleted
to increase the likelihood that it is used again for future models.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>