feat: Support for downloading logs
Accessing the browser console log is unfamiliar to many
non-technical users, so this commit adds a new feature to do
this through the UI. It's hidden unless someone adds "?debug"
to the URL, which I'll do when someone raises a support issue.

Also includes an update in Node.js to 12.x, and a variety of
npm dependency updates.

The dependency updates required a few code changes, because of
default behaviour changes in Node.js (v12), Gulp (v4)
and Helmet.

Closes: #326

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>