chore: Host dependencies with the site instead of using CDN
TensorflowJS and D3 were being loaded from unpkg.com but are now
hosted with the rest of dependencies fetched by bower.

I've also updated to a newer version of tfjs, and fixed a minor
bug that caused console errors if you click away from the models
page while leaving the model running for a test.

Closes: #350

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>