feat: Prototype for training image classifiers in the browser
This commit introduces a new project type ("imgtfjs") which is
functionally identical to "images" except doesn't have the
code for creating/using/managing Watson Visual Recognition models.

This is a start to allow TensorFlowJS support to be developed in
parallel with the existing site, as it means I don't have to
modify any of the image project functionality until I'm ready.

I've added a new API endpoint for allowing the browser to download
training images (this is difficult to do from the browser because
of CORS errors when TFJS tries to use them). This means the client
side code can assume that the images are already resized and
scaled before starting to train the model.

Contributes to: #351

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>