feat: Support for training image classifiers in the browser (#358)
This commit introduces a new project type ("imgtfjs") which is
functionally identical to "images" except doesn't have the
code for creating/using/managing Watson Visual Recognition models.

Instead of using Watson APIs, models for imgtjfs projects are 
trained in the browser using TensorFlow.js. 

I've added a new API endpoint for allowing the browser to download
training images (this is difficult to do from the browser because
of CORS errors when TFJS tries to use them). This means the client
side code can assume that the images are already resized and
scaled before starting to train the model.

To reduce the overhead of all of this, the image download APIs have
aggressive cache headers set, and models are stored in browser 
storage using indexeddb to allow them to be reused.

This needs more testing to confirm it will work correctly on the
types of projects that students build using the site, and on the 
kinds of computers that schools and code clubs use. Because of
this, the new image functionality is hidden at both the API level 
(usable only by classes with 'imgtfjs' in their supported projects
types list) and in the UI (shown only if a beta url variable is added).

Contributes to: #351

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>