feat: Enable in-browser TensorFlow.js training for all users
This adds the final missing feature to enable parity between
WVR and TFJS training which is to test an image classifier by
providing a URL. To avoid CORS issues, retrieving the image
data is deferred to the server, which can then be tested in the
browser.

With this final feature, and a week's successful testing of the
private beta, I'm now making this widely available. Image
projects can now be trained in locally in the browser by all
classes.

I'm not removing Watson Visual Recognition at this point, as
a) the TensorFlow.js support hasn't been tested widely enough yet
b) TensorFlow support doesn't work for App Inventor or Python yet

This commit enables the feature, as well as allowing students
to switch between the two project types, with some additional
documentation to explain the difference between in-browser and
Watson Visual Recognition training.

Closes: #351

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>