feat: Notify students if the add-training-data blocks fail
If the limit for the number of training examples is reached, the
add-training-data blocks in Scratch fail silently - you can only
tell if you know to check the browser's network tab to see the
http-409 responses.

This commit adds a new help screen that pops-up after a http-409
response to notify students that no new examples can be added.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>