fix: avoid rate-limit errors in Python projects
Fix to handle reports of rate-limiting errors that students were
getting on Python image projects using training images from
Wikipedia.

Wikipedia applies agressive rate-limiting policies for requests
that don't identify themselves with a user-agent header.
The TensorFlow/Keras utility that downloads training items doesn't
include a user-agent in the requests that it makes. To avoid
needing to make changes to those libraries, this commit instead
installs a global request handler to add a user-agent header.

I've also added a brief sleep as well, to conform with the
recommendation from the wikipedia robot policy at
https://wikitech.wikimedia.org/wiki/Robot_policy
As training images are cached after the first download, this
shouldn't have a big impact on the training time for student
projects.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>