fix: handle auth/login failures caused by ipv6 errors
Requests forwarded to auth0 by login.machinelearningforkids.co.uk are
failing. The nginx log shows errors like

connect() to [2a00:1450:4001:824::2010]:443 failed (101: Network unreachable) while connecting to upstream
upstream server temporarily disabled while connecting to upstream

I *think* this is because the auth0 hostname that we are using
proxy_pass to forward to is being resolved to an ipv6 address.

I've tried a few approaches to disable ipv6, but nothing has been
successful. In this commit, I'm giving up and resorting to using an old
version as described in
https://github.com/nginxinc/docker-nginx/issues/374

While I'm not super comfortable using an older version of nginx, I
can't find another way that works.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>